Renamed and changed the demo/train-cifar10-torch so it's importable.
