+++ demo/train_cifar10_tf/data.py
