+++ demo/train_cifar10_tf/model.py
