+++ demo/train_cifar10_tf/train.py
