+++ demo/train_cifar10_torch/model.py
